Transformed Health and Attack Power to Integer (was Float)

+++ src/main/java/org/davidfabio/game/Attacker.java
@@ -3,7 +3,7 @@ package org.davidfabio.game;
 import com.badlogic.gdx.graphics.Color;
 
 public interface Attacker {
-    public float getAttackPower();
+    public int getAttackPower();
 
     public default void attack(Attackable attackable, World world) {
         if (!attackable.getIsActive())
